Added confirmation alert for deleting things

+++ frontend/public/scripts/postElement.js
@@ -1,6 +1,6 @@
 // this function gets reused by multiple files
 
-function createPostElement(id, contents, username, timestamp, link=false) {
+function createPostElement(id, contents, username, timestamp, viewerUsername, link=false) {
     const div = document.createElement('div');
 
     const usernameElement = document.createElement('div');
@@ -56,6 +56,9 @@ function createPostElement(id, contents, username, timestamp, link=false) {
     const secondDiv = document.createElement('div');
 
     deleteButton.addEventListener('click', async () => {
+        if (username !== viewerUsername) {return;}
+        let confirmDelete = confirm('Are you sure you want to delete this post?');
+        if (!confirmDelete) {return;}
         const response = await fetch(`/api/posts/${id}`, {
             method: 'DELETE',
         });
